+++ src/main/java/com/epam/ta/reportportal/core/item/impl/status/ToSkippedStatusChangingStrategy.java
@@ -35,6 +35,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 import static com.epam.ta.reportportal.commons.Preconditions.statusIn;
@@ -70,39 +71,42 @@ public class ToSkippedStatusChangingStrategy extends AbstractStatusChangingStrat
 				);
 
 		testItem.getItemResults().setStatus(providedStatus);
-		Optional<ItemAttribute> skippedIssueAttribute = itemAttributeRepository.findByLaunchIdAndKeyAndSystem(testItem.getLaunchId(),
-				SKIPPED_ISSUE_KEY,
-				true
-		);
 
-		boolean issueRequired = skippedIssueAttribute.isPresent() && BooleanUtils.toBoolean(skippedIssueAttribute.get().getValue());
+		if (Objects.isNull(testItem.getRetryOf())) {
+			Optional<ItemAttribute> skippedIssueAttribute = itemAttributeRepository.findByLaunchIdAndKeyAndSystem(testItem.getLaunchId(),
+					SKIPPED_ISSUE_KEY,
+					true
+			);
 
-		if (issueRequired) {
-			if (testItem.getItemResults().getIssue() == null && testItem.isHasStats()) {
-				addToInvestigateIssue(testItem, project.getId());
+			boolean issueRequired = skippedIssueAttribute.isPresent() && BooleanUtils.toBoolean(skippedIssueAttribute.get().getValue());
+
+			if (issueRequired) {
+				if (testItem.getItemResults().getIssue() == null && testItem.isHasStats()) {
+					addToInvestigateIssue(testItem, project.getId());
+				}
+			} else {
+				ofNullable(testItem.getItemResults().getIssue()).map(issue -> {
+					issue.setTestItemResults(null);
+					testItem.getItemResults().setIssue(null);
+					return issue.getIssueId();
+				}).ifPresent(issueEntityRepository::deleteById);
 			}
-		} else {
-			ofNullable(testItem.getItemResults().getIssue()).map(issue -> {
-				issue.setTestItemResults(null);
-				testItem.getItemResults().setIssue(null);
-				return issue.getIssueId();
-			}).ifPresent(issueEntityRepository::deleteById);
+
+			List<Long> itemsToReindex = changeParentsStatuses(testItem, launch, true, user);
+			itemsToReindex.add(testItem.getItemId());
+			logIndexer.cleanIndex(project.getId(),
+					logRepository.findIdsUnderTestItemByLaunchIdAndTestItemIdsAndLogLevelGte(testItem.getLaunchId(),
+							itemsToReindex,
+							LogLevel.ERROR.toInt()
+					)
+			);
+
+			if (!issueRequired) {
+				itemsToReindex.remove(itemsToReindex.size() - 1);
+			}
+
+			logIndexer.indexItemsLogs(project.getId(), launch.getId(), itemsToReindex, AnalyzerUtils.getAnalyzerConfig(project));
 		}
-
-		List<Long> itemsToReindex = changeParentsStatuses(testItem, launch, true, user);
-		itemsToReindex.add(testItem.getItemId());
-		logIndexer.cleanIndex(project.getId(),
-				logRepository.findIdsUnderTestItemByLaunchIdAndTestItemIdsAndLogLevelGte(testItem.getLaunchId(),
-						itemsToReindex,
-						LogLevel.ERROR.toInt()
-				)
-		);
-
-		if (!issueRequired) {
-			itemsToReindex.remove(itemsToReindex.size() - 1);
-		}
-
-		logIndexer.indexItemsLogs(project.getId(), launch.getId(), itemsToReindex, AnalyzerUtils.getAnalyzerConfig(project));
 	}
 
 	@Override
